move mojang & namespace tests
