+++ src/test/java/net/minestom/server/utils/TestMojangUtils.java
@@ -1,4 +1,4 @@
-package misc;
+package net.minestom.server.utils;
 
 import net.minestom.server.utils.mojang.MojangUtils;
 import org.junit.jupiter.api.Test;
